requirements.txt
requests
selenium
PyJWT
PyYAML
fastapi
aiofiles
python-multipart
uvicorn
httpx